code/backend/pyproject.toml
[tool.poetry]
name = "project-noe"
version = "0.1.0"
description = "Project Noe is founded to help people and government(s) to prevent the spread of COVID19."
authors = ["Rollet Developers"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.5"
djangorestframework = "^3.11.0"
requests = "^2.23.0"
django-cors-headers = "^3.2.1"
python-dotenv = "^0.12.0"
cryptography = "^2.9"
django-filter = "^2.2.0"
qrcode = {version = "^6.1", extras = ["pil"]}
psycopg2-binary = "^2.8.5"
phonenumbers = "^8.12.1"
unidecode = "^1.1.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest = "^5.4.1"
pytest-django = "^3.9.0"

[tool.black]
line-length = 119
exclude = "migrations/.*"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

